+++ electron-builder.yml
@@ -1,6 +1,6 @@
 appId: io.settlenet.signtool
 productName: LedgerSignTool
-copyright: "Copyright © 2020 Crypto Garage, Inc. All Rights Reserved."
+copyright: "Copyright © 2021 Crypto Garage, Inc. All Rights Reserved."
 
 directories:
   output: dist
@@ -10,7 +10,7 @@ files:
   - filter:
       - ./main.js
       - ./renderer.js
-      - ./txFileUtil.js
+      - ./preload.js
       - ./index.html
 
 extraMetadata:
@@ -18,30 +18,30 @@ extraMetadata:
 
 compression: normal
 
-#afterSign: "dev_script/package/notarize/notarize.js"
+afterSign: "dev_script/package/notarize/notarize.js"
 
 mac:
   target:
     - zip
-    - dmg
-#  hardenedRuntime: true
-#  gatekeeperAssess: false
+#    - dmg
+  hardenedRuntime: true
+  gatekeeperAssess: false
 #  entitlements: "dev_script/package/notarize/entitlements.mac.plist"
 #  entitlementsInherit: "dev_script/package/notarize/entitlements.mac.plist"
-  publish:
-    provider: "github"
-    private: true
 
 win:
   target:
     - zip
     - nsis
-  publish:
-    provider: "github"
-    private: true
 
 nsis:
   createDesktopShortcut: true
   createStartMenuShortcut: true
+
+publish:
+  provider: "github"
+  releaseType: "release"
+  vPrefixedTagName: true
+  owner: "cryptogarageinc"
 
 extends: null
