+++ index.html
@@ -110,9 +110,7 @@
       </div>
     </div>
 
-    <script>
-      // You can also require other files to run in this process
-      require("./renderer.js");
+    <script type="text/javascript" src="renderer.js">
     </script>
   </body>
 </html>
